fix： 修改圈子的数据结构

+++ src/main/java/com/idreamsky/fanbook/sdk/bot/model/v20220429/CircleCover.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class CircleCover {
 
-    private Integer height;
+    private Double height;
 
     private String source;
 
-    private Integer width;
+    private Double width;
 }
